refactor(document): split pptx writer responsibilities
