+++ packages/document/src/writer/part-serializers.ts
@@ -1,0 +1,91 @@
+import type {
+  PartRelationships,
+  PptxSourceModel,
+  RawOoxmlNode,
+  RawPackagePart,
+} from "../source/index.js";
+import { encodeXml, XML_DECLARATION } from "./xml-serialization.js";
+
+export function serializeContentTypes(
+  contentTypes: PptxSourceModel["packageGraph"]["contentTypes"],
+): string {
+  const defaults = contentTypes.defaults
+    .map(
+      (entry) =>
+        `<Default Extension="${escapeAttribute(entry.extension)}" ` +
+        `ContentType="${escapeAttribute(entry.contentType)}"/>`,
+    )
+    .join("");
+  const overrides = contentTypes.overrides
+    .map(
+      (entry) =>
+        `<Override PartName="/${escapeAttribute(entry.partName)}" ` +
+        `ContentType="${escapeAttribute(entry.contentType)}"/>`,
+    )
+    .join("");
+  return (
+    XML_DECLARATION +
+    `<Types xmlns="http://schemas.openxmlformats.org/package/2006/content-types">` +
+    defaults +
+    overrides +
+    `</Types>`
+  );
+}
+
+export function serializeRelationships(partRelationships: PartRelationships): string {
+  const relationships = partRelationships.relationships
+    .map((relationship) => {
+      const targetMode =
+        relationship.targetMode === undefined
+          ? ""
+          : ` TargetMode="${escapeAttribute(relationship.targetMode)}"`;
+      return (
+        `<Relationship Id="${escapeAttribute(relationship.id)}" ` +
+        `Type="${escapeAttribute(relationship.type)}" ` +
+        `Target="${escapeAttribute(relationship.target)}"${targetMode}/>`
+      );
+    })
+    .join("");
+  return (
+    XML_DECLARATION +
+    `<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">` +
+    relationships +
+    `</Relationships>`
+  );
+}
+
+export function serializeRawPackagePart(rawPart: RawPackagePart): Uint8Array {
+  if (rawPart.kind === "binary") return rawPart.bytes;
+  return encodeXml(XML_DECLARATION + serializeRawNode(rawPart.xml));
+}
+
+export function serializeRawNode(node: RawOoxmlNode): string {
+  const attributes =
+    node.attributes === undefined
+      ? ""
+      : Object.entries(node.attributes)
+          .map(([name, value]) => ` ${name}="${escapeAttribute(value)}"`)
+          .join("");
+  const text = node.text === undefined ? "" : escapeText(node.text);
+  const children = node.children?.map((child) => serializeRawNode(child)).join("") ?? "";
+  if (text !== "" && children !== "") {
+    throw new Error(
+      `writePptx: raw XML part '${node.name}' contains mixed text/element content; ` +
+        "ordered mixed-content serialization is not implemented in the no-edit writer",
+    );
+  }
+  if (text === "" && children === "") return `<${node.name}${attributes}/>`;
+  return `<${node.name}${attributes}>${text}${children}</${node.name}>`;
+}
+
+function escapeAttribute(value: string): string {
+  return value
+    .replaceAll("&", "&amp;")
+    .replaceAll('"', "&quot;")
+    .replaceAll("<", "&lt;")
+    .replaceAll(">", "&gt;");
+}
+
+function escapeText(value: string): string {
+  return value.replaceAll("&", "&amp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;");
+}
